test: ethProxy

+++ contracts/IUSM.sol
@@ -4,9 +4,9 @@ pragma solidity ^0.6.7;
 interface IUSM {
     enum Side {Buy, Sell}
 
-    function mint(address from, address to, uint256 ethIn) external returns (uint);
+    function mint(address from, address to, uint ethIn) external returns (uint);
     function burn(address from, address to, uint usmToBurn) external returns (uint);
-    function fund(address from, address to, uint256 ethIn) external returns (uint);
+    function fund(address from, address to, uint ethIn) external returns (uint);
     function defund(address from, address to, uint fumToBurn) external returns (uint);
     function ethPool() external view returns (uint);
     function ethBuffer() external view returns (int);
